use RatingMixin.ts for all ratings components
